+++ index.d.ts
@@ -1,3 +1,0 @@
-﻿export * from './exemplify/services/source.service';
-export * from './exemplify/directives/add-example.directive';
-export * from './exemplify/exemplify.module';