+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Release History
 ---------------
 ## [Unreleased]
+# [1.0.0] - 2017-01-03
+
+### Added
+- Added show/hide toggle function
+
+### Changed
+- Behaviour for nested elements, ie. if element becomes nested due to `*ngIf` or other directives
+
 # [1.0.0-rc.1] - 2017-01-03
 
 ### Added